requirements.txt
streamlit
requests
python-dotenv
markdown
pyyaml